requirements.txt
art == 6.0
requests == 2.31.0
periodictable == 1.6.1
azure-identity
azure-keyvault-secrets